% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FHSTR-package.R
\docType{package}
\name{FHSTR-package}
\alias{FHSTR}
\alias{FHSTR-package}
\title{FHSTR: 2022 Beijing Olympic Data from NBC's API}
\description{
This package currently provides data from the 2022 Beijing Olympics from NBC's API. The name of this package comes from the Olympic Motto: "Faster, Higher, Stronger - Together". This package is still a work in progress and new functions will hopefully be added soon! This package will likely be updated to include Paris 2024 Data when that event comes around.
}
\author{
\strong{Maintainer}: Billy Fryer \email{billyfryer@att.net}

}
\keyword{internal}
