% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_olympic_csv_data}
\alias{load_olympic_csv_data}
\title{Load Olympic CSV Data}
\source{
NBC API
}
\usage{
load_olympic_csv_data(sportID, matchID)
}
\arguments{
\item{sportID}{Sport ID}

\item{matchID}{Match ID}
}
\value{
A data.frame
}
\description{
Given a Sport ID and Match ID, load a pre-parsed CSV of data from that event
}
\examples{
\donttest{
load_olympic_csv_data(sportID = 113, matchID = 746587)
}
}
\author{
Billy Fryer
}
