% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_olympic_json_data}
\alias{load_olympic_json_data}
\title{Load Olympic JSON Data}
\source{
NBC API
}
\usage{
load_olympic_json_data(sportID, matchID)
}
\arguments{
\item{sportID}{Sport ID}

\item{matchID}{Match ID}
}
\value{
Nested Lists in the form of a JSON File
}
\description{
Given a Sport ID and Match ID, load the raw JSON file
}
\examples{
\donttest{
load_olympic_json_data(sportID = 113, matchID = 746587)
}
}
\author{
Billy Fryer
}
