% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_others.R
\name{load_olympic_matchID_key}
\alias{load_olympic_matchID_key}
\title{Load Olympic MatchID Keys}
\source{
NBC API
}
\usage{
load_olympic_matchID_key(sportID)
}
\arguments{
\item{sportID}{Sport ID}
}
\value{
A data.frame
}
\description{
Given a sportID, list all possible matchIDs as well as the corresponding event
}
\examples{
\donttest{
load_olympic_matchID_key(sportID = 113)
}
}
\author{
Billy Fryer
}
