% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_others.R
\name{load_olympic_sport_schedules}
\alias{load_olympic_sport_schedules}
\title{Load Olympic Sport Schedules}
\source{
NBC API
}
\usage{
load_olympic_sport_schedules(sportID)
}
\arguments{
\item{sportID}{Spor ID}
}
\value{
A data.frame
}
\description{
Given a sportID, list all Sport Schedules
}
\examples{
\donttest{
try(load_olympic_sport_schedules(sportID = 113))
}
}
\author{
Billy Fryer
}
