% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_matrix.R
\docType{data}
\name{schedule_matrix}
\alias{schedule_matrix}
\title{Schedule Matrix}
\format{
A data.frame with 285 rows and 16 variables
\describe{
  \item{SportID}{int Respective Sport ID for each sport}
  \item{Sport}{chr Name of Sport}
  \item{SportShort}{chr Abbreviation for Sport}
  \item{n_DateLocal}{int Date in YYYYMMDD format}
  \item{c_DateLoca}{POSIXCT Date in POSIXCT format}
  \item{d_DateLocal}{chr Date in UNIX format}
  \item{TimeFirstEventLocal}{POSIXCT Start time in Beijing in POSIXCT format}
  \item{TimeFirstEventUTC}{POSIXCT Start time in UTC in POSIXCT format}
  \item{TimeLastEventLocal}{POSIXCT Last time in Beijing in POSIXCT format}
  \item{TimeLastEventUTC}{POSIXCT Last time in UTC in POSIXCT format}
  \item{NrOfEvents}{int Number of Events that Day}
  \item{NrOfMatches}{int Number of Matches that Day}
  \item{NrOfGold}{int Number of Gold Medals Awarded that Day}
  \item{NrOfSilver}{int Number of Silver Medals Awarded that Day}
  \item{NrOfBronze}{int Number of Bronze Medals Awarded that Day}
}
}
\source{
NBC Olympics API
}
\usage{
schedule_matrix
}
\description{
Data about Event Start Times during the 2022 Beijing Olympics
}
\keyword{datasets}
