% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_matrix.R
\docType{data}
\name{schedule_matrix}
\alias{schedule_matrix}
\title{Schedule Matrix}
\format{
A data.frame with 285 rows and 16 variables
\describe{
  \item{n_SportID}{int Respective Sport ID for each sport}
  \item{c_Sport}{chr Name of Sport}
  \item{c_SportShort}{chr Abbreviation for Sport}
  \item{n_DateLocal}{int Date in YYYYMMDD format}
  \item{c_DateLoca}{POSIXCT Date in POSIXCT format}
  \item{d_DateLocal}{chr Date in UNIX format}
  \item{c_TimeFirstEventLocal}{POSIXCT Start time in Beijing in POSIXCT format}
  \item{c_TimeFirstEventUTC}{POSIXCT Start time in UTC in POSIXCT format}
  \item{c_TimeLastEventLocal}{POSIXCT Last time in Beijing in POSIXCT format}
  \item{c_TimeLastEventUTC}{POSIXCT Last time in UTC in POSIXCT format}
  \item{n_NrOfEvents}{int Number of Events}
  \item{n_NrOfPhases}{int Number of Phases}
  \item{n_NrOfMatches}{int Number of Matches}
  \item{n_NrOfGold}{int Number of Gold Medals}
  \item{n_NrOfSilver}{int Number of Silver Medals}
  \item{n_NrOfBronze}{int Number of Bronze Medals}
}
}
\source{
NBC Olympics API
}
\usage{
schedule_matrix
}
\description{
Data about Event Start Times during the 2022 Beijing Olympics
}
\keyword{datasets}
